NOTE-TEKAR
The NoteTaker application is a web-based note-taking application developed using Bootstrap for
responsive and visually , HTML for structuring the web pages, Java Servlets for server-side processing,
and Hibernate for connecting to and managing the database. The application allows users to create,
view, update, and delete notes And database quary automaticaly generat using Hibernate.